[1.8.4] AntiTrap - Bomb, HighFacePlace, Refactoring, #241
