+++ src/main/java/me/earth/earthhack/impl/modules/combat/antitrap/AntiTrap.java
@@ -1,18 +1,23 @@
 package me.earth.earthhack.impl.modules.combat.antitrap;
 
+import me.earth.earthhack.api.cache.ModuleCache;
 import me.earth.earthhack.api.module.util.Category;
 import me.earth.earthhack.api.setting.Complexity;
 import me.earth.earthhack.api.setting.Setting;
 import me.earth.earthhack.api.setting.settings.BooleanSetting;
 import me.earth.earthhack.api.setting.settings.EnumSetting;
 import me.earth.earthhack.api.setting.settings.NumberSetting;
+import me.earth.earthhack.impl.modules.Caches;
 import me.earth.earthhack.impl.modules.combat.antitrap.util.AntiTrapMode;
+import me.earth.earthhack.impl.modules.combat.offhand.Offhand;
 import me.earth.earthhack.impl.modules.combat.offhand.modes.OffhandMode;
-import me.earth.earthhack.impl.util.helpers.blocks.ObbyModule;
+import me.earth.earthhack.impl.util.helpers.blocks.ObbyListenerModule;
 import me.earth.earthhack.impl.util.math.StopWatch;
 import me.earth.earthhack.impl.util.math.position.PositionUtil;
+import me.earth.earthhack.impl.util.minecraft.InventoryUtil;
 import me.earth.earthhack.impl.util.minecraft.blocks.BlockUtil;
 import net.minecraft.entity.item.EntityEnderCrystal;
+import net.minecraft.init.Blocks;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.RayTraceResult;
@@ -21,8 +26,11 @@ import net.minecraft.util.math.Vec3i;
 import java.util.*;
 
 // TODO: INSTANT!!!!
-public class AntiTrap extends ObbyModule
+public class AntiTrap extends ObbyListenerModule<ListenerAntiTrap>
 {
+    private static final ModuleCache<Offhand> OFFHAND =
+        Caches.getModule(Offhand.class);
+
     protected final Setting<AntiTrapMode> mode =
         registerBefore(new EnumSetting<>("Mode", AntiTrapMode.Crystal), blocks);
     protected final Setting<Boolean> offhand =
@@ -40,11 +48,13 @@ public class AntiTrap extends ObbyModule
     protected final Setting<Boolean> highFill =
         register(new BooleanSetting("HighFill", false))
             .setComplexity(Complexity.Medium);
-    protected final Setting<Integer> confirm =
-        register(new NumberSetting<>("Confirm", 250, 0, 1000))
-            .setComplexity(Complexity.Expert);
+    protected final Setting<Boolean> highFacePlace =
+        register(new BooleanSetting("HighFacePlace", false))
+            .setComplexity(Complexity.Medium);
     protected final Setting<Boolean> autoOff =
         register(new BooleanSetting("Auto-Off", true));
+    protected final Setting<Boolean> requireOffhand =
+        register(new BooleanSetting("RequireOffhand", false));
 
     /** Blocks that have been placed an await a SPacketBlockChange */
     protected final Map<BlockPos, Long> placed = new HashMap<>();
@@ -60,7 +70,6 @@ public class AntiTrap extends ObbyModule
     public AntiTrap()
     {
         super("AntiTrap", Category.Combat);
-        this.listeners.add(new ListenerMotion(this));
         this.setData(new AntiTrapData(this));
     }
 
@@ -95,8 +104,48 @@ public class AntiTrap extends ObbyModule
     {
         if (offhand.getValue() && previous != null)
         {
-            ListenerMotion.OFFHAND.computeIfPresent(o -> o.setMode(previous));
+            OFFHAND.computeIfPresent(o -> o.setMode(previous));
         }
+    }
+
+    @Override
+    protected ListenerAntiTrap createListener()
+    {
+        return new ListenerAntiTrap(this);
+    }
+
+    @Override
+    public boolean execute()
+    {
+        if (offhand.getValue()
+            && this.listener.mode != AntiTrapMode.Crystal
+            && OFFHAND.isEnabled())
+        {
+            OffhandMode previous = this.previous;
+            if (packets.isEmpty()
+                && previous != null
+                && OFFHAND.get().getMode() == OffhandMode.OBSIDIAN)
+            {
+                OFFHAND.get().setMode(previous);
+                this.previous = null;
+            }
+            else if (!packets.isEmpty()
+                && OFFHAND.get().getMode() != OffhandMode.OBSIDIAN
+                && OFFHAND.get().isSafe())
+            {
+                this.previous = OFFHAND.get().getMode();
+                OFFHAND.get().setMode(OffhandMode.OBSIDIAN);
+                OFFHAND.get().doOffhand();
+            }
+
+            if (requireOffhand.getValue()
+                && !InventoryUtil.isHolding(Blocks.OBSIDIAN))
+            {
+                return false;
+            }
+        }
+
+        return super.execute();
     }
 
     @Override
